+++ src/FastAccelStepper.h
@@ -721,11 +721,11 @@ class FastAccelStepper {
   // of the GPIOs, then the additional dir_pin_readback parameter informs
   // about this pin.
   //
-  // Update for idf5 version:
+  // Update for idf5/idf6 version:
   // The pcnt_unit value is not used, because the available units are managed
   // by the system. The parameter is kept for compatibility.
   //
-#if defined(SUPPORT_ESP32_PULSE_COUNTER) && (ESP_IDF_VERSION_MAJOR == 5)
+#if defined(SUPPORT_ESP32_PULSE_COUNTER) && (ESP_IDF_VERSION_MAJOR >= 5)
   bool attachToPulseCounter(uint8_t unused_pcnt_unit = 0,
                             int16_t low_value = -16384,
                             int16_t high_value = 16384,
@@ -783,7 +783,7 @@ class FastAccelStepper {
 
   ExtDirPendingState _pendingExternalDirState;
 
-#if defined(SUPPORT_ESP32_PULSE_COUNTER) && (ESP_IDF_VERSION_MAJOR == 5)
+#if defined(SUPPORT_ESP32_PULSE_COUNTER) && (ESP_IDF_VERSION_MAJOR >= 5)
   pcnt_unit_handle_t _attached_pulse_unit;
 #endif
 #if defined(SUPPORT_ESP32_PULSE_COUNTER) && (ESP_IDF_VERSION_MAJOR == 4)