+++ src/pd_esp32/pd_config_idf6.h
@@ -11,7 +11,7 @@
 // #define SUPPORT_ESP32_MCPWM_PCNT
 #define SUPPORT_ESP32_RMT
 #define SUPPORT_ESP32_RMT_V2
-// #define SUPPORT_ESP32_PULSE_COUNTER 8
+#define SUPPORT_ESP32_PULSE_COUNTER 8
 #define HAVE_ESP32_RMT
 #define RMT_SIZE 64
 
@@ -21,7 +21,7 @@
 
 #define NEED_RMT_HEADERS
 // #define NEED_MCPWM_HEADERS
-// #define NEED_PCNT_HEADERS
+#define NEED_PCNT_HEADERS
 
 //==========================================================================
 //
@@ -31,14 +31,14 @@
 #elif CONFIG_IDF_TARGET_ESP32S2
 #define SUPPORT_ESP32_RMT
 #define SUPPORT_ESP32_RMT_V2
-// #define SUPPORT_ESP32_PULSE_COUNTER 4
+#define SUPPORT_ESP32_PULSE_COUNTER 4
 // #define HAVE_ESP32S3_PULSE_COUNTER
 #define HAVE_ESP32_RMT
 #define RMT_SIZE 64
 #define QUEUES_MCPWM_PCNT 0
 #define QUEUES_RMT 4
 #define NEED_RMT_HEADERS
-// #define NEED_PCNT_HEADERS
+#define NEED_PCNT_HEADERS
 
 //==========================================================================
 //
@@ -49,7 +49,7 @@
 // #define SUPPORT_ESP32_MCPWM_PCNT
 #define SUPPORT_ESP32_RMT
 #define SUPPORT_ESP32_RMT_V2
-// #define SUPPORT_ESP32_PULSE_COUNTER 8
+#define SUPPORT_ESP32_PULSE_COUNTER 8
 // #define HAVE_ESP32S3_PULSE_COUNTER
 #define HAVE_ESP32_RMT
 #define RMT_SIZE 48
@@ -59,7 +59,7 @@
 #define QUEUES_RMT 4
 #define NEED_RMT_HEADERS
 // #define NEED_MCPWM_HEADERS
-// #define NEED_PCNT_HEADERS
+#define NEED_PCNT_HEADERS
 
 //==========================================================================
 //
@@ -83,13 +83,13 @@
 #elif CONFIG_IDF_TARGET_ESP32C6
 #define SUPPORT_ESP32_RMT
 #define SUPPORT_ESP32_RMT_V2
-// #define SUPPORT_ESP32_PULSE_COUNTER 4
+#define SUPPORT_ESP32_PULSE_COUNTER 4
 #define HAVE_ESP32_RMT
 #define RMT_SIZE 48
 #define QUEUES_MCPWM_PCNT 0
 #define QUEUES_RMT 2
 #define NEED_RMT_HEADERS
-// #define NEED_PCNT_HEADERS
+#define NEED_PCNT_HEADERS
 
 //==========================================================================
 //
@@ -99,13 +99,13 @@
 #elif CONFIG_IDF_TARGET_ESP32H2
 #define SUPPORT_ESP32_RMT
 #define SUPPORT_ESP32_RMT_V2
-// #define SUPPORT_ESP32_PULSE_COUNTER 4
+#define SUPPORT_ESP32_PULSE_COUNTER 4
 #define HAVE_ESP32_RMT
 #define RMT_SIZE 48
 #define QUEUES_MCPWM_PCNT 0
 #define QUEUES_RMT 2
 #define NEED_RMT_HEADERS
-// #define NEED_PCNT_HEADERS
+#define NEED_PCNT_HEADERS
 
 //==========================================================================
 //
@@ -115,13 +115,13 @@
 #elif CONFIG_IDF_TARGET_ESP32P4
 #define SUPPORT_ESP32_RMT
 #define SUPPORT_ESP32_RMT_V2
-// #define SUPPORT_ESP32_PULSE_COUNTER CONFIG_SOC_PCNT_UNITS_PER_GROUP
+#define SUPPORT_ESP32_PULSE_COUNTER CONFIG_SOC_PCNT_UNITS_PER_GROUP
 #define HAVE_ESP32_RMT
 #define RMT_SIZE CONFIG_SOC_RMT_MEM_WORDS_PER_CHANNEL
 #define QUEUES_MCPWM_PCNT 0
 #define QUEUES_RMT CONFIG_SOC_RMT_TX_CANDIDATES_PER_GROUP
 #define NEED_RMT_HEADERS
-// #define NEED_PCNT_HEADERS
+#define NEED_PCNT_HEADERS
 
 //==========================================================================
 //
@@ -144,11 +144,8 @@
 
 #ifdef NEED_PCNT_HEADERS
 #include <driver/pulse_cnt.h>
-// #include <soc/pcnt_reg.h>
-// #include <soc/pcnt_struct.h>
-#include <soc/pcnt_periph.h>
+#include <hal/pcnt_periph.h>
 #include <driver/gpio.h>
-#include <rom/gpio.h>
 #include <hal/gpio_ll.h>
 #include <esp_rom_gpio.h>
 #endif